test/s3_data_aggregator/requirements.txt
aws-cdk-lib
constructs
types-boto3
types-setuptools
boto3
aws-cdk.aws-kinesisfirehose-alpha
aws-cdk.aws-kinesisfirehose-destinations-alpha